# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/resource-donations": {
      "get": {
        "operationId": "ResourceDonationsController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource donations retrieved"
          },
          "403": {
            "description": "Access denied — requires admin or superuser role"
          }
        },
        "summary": "Get all resource donations",
        "tags": [
          "Resource Donations"
        ]
      },
      "post": {
        "description": "Donor submits a resource donation with pickup details",
        "operationId": "ResourceDonationsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDonationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource donation submitted successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "403": {
            "description": "Access denied — requires donor, admin, or superuser role"
          }
        },
        "summary": "Create a new resource donation",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/pending": {
      "get": {
        "operationId": "ResourceDonationsController_getPending",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending donations retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get pending (submitted) donations",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/active-pickups": {
      "get": {
        "operationId": "ResourceDonationsController_getActivePickups",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active pickups retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get active pickups assigned to volunteers",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/by-donor/{email}": {
      "get": {
        "operationId": "ResourceDonationsController_findByDonor",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Donations retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get donations by donor email",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/by-volunteer/{email}": {
      "get": {
        "operationId": "ResourceDonationsController_findByVolunteer",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Donations retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get donations assigned to volunteer for pickup",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/by-program/{programId}": {
      "get": {
        "operationId": "ResourceDonationsController_findByProgram",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Donations retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get donations allocated to a program",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/by-status/{status}": {
      "get": {
        "operationId": "ResourceDonationsController_findByStatus",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Donations retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get donations by status",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/{id}": {
      "get": {
        "operationId": "ResourceDonationsController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Donation retrieved"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "summary": "Get resource donation by ID",
        "tags": [
          "Resource Donations"
        ]
      },
      "patch": {
        "operationId": "ResourceDonationsController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDonationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Donation updated"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "summary": "Update a resource donation",
        "tags": [
          "Resource Donations"
        ]
      },
      "delete": {
        "operationId": "ResourceDonationsController_remove",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Donation deleted"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "summary": "Delete a resource donation",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/{id}/assign-volunteer": {
      "patch": {
        "operationId": "ResourceDonationsController_assignVolunteer",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Volunteer assigned for pickup"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "summary": "Assign a volunteer for pickup",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/{id}/start-pickup": {
      "patch": {
        "operationId": "ResourceDonationsController_startPickup",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pickup started"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "summary": "Start pickup process",
        "tags": [
          "Resource Donations"
        ]
      }
    },
...
    "/resource-donations/{id}/mark-completed": {
      "patch": {
        "operationId": "ResourceDonationsController_markCompleted",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marked as completed"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "summary": "Mark donation as completed",
        "tags": [
          "Resource Donations"
        ]
      }
    },
    "/resource-donations/{id}/allocate-program": {
      "patch": {
        "operationId": "ResourceDonationsController_allocateToProgram",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Allocated to program"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "summary": "Allocate donation to a program",
        "tags": [
          "Resource Donations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kindred Platform API",
    "description": "Backend REST API for Kindred — India's Volunteer Platform. Supports Users, Organizations, Volunteers, Programs, Requests, Tasks, Hour Logs, Resource Donations, Program Assignments, Program Applications, and Subtasks. Role-based access control via the \"role\" header.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "DonationItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Item name",
            "example": "Rice Bags (25kg)"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity",
            "example": 10
          },
          "condition": {
            "type": "string",
            "description": "Condition of items",
            "example": "New"
          }
        },
        "required": [
          "name",
          "quantity",
          "condition"
        ]
      },
      "CreateResourceDonationDto": {
        "type": "object",
        "properties": {
          "donorName": {
            "type": "string",
            "description": "Donor full name",
            "example": "Vikram Malhotra"
          },
          "donorEmail": {
            "type": "string",
            "description": "Donor email",
            "example": "vikram.malhotra@gmail.com"
          },
          "donorPhone": {
            "type": "string",
            "description": "Donor phone",
            "example": "9812345678"
          },
          "category": {
            "type": "string",
            "description": "Category of donation",
            "example": "Food & Essentials"
          },
          "items": {
            "description": "List of donated items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DonationItemDto"
            }
          },
          "description": {
            "type": "string",
            "description": "Brief description",
            "example": "Bulk food supplies for flood relief camps in Bihar"
          },
          "pickupAddress": {
            "type": "string",
            "description": "Pickup address",
            "example": "42, MG Road, Sector 18, Noida, UP"
          },
          "pickupDate": {
            "type": "string",
            "description": "Pickup date (YYYY-MM-DD)",
            "example": "2025-05-10"
          },
          "pickupTimeSlot": {
            "type": "string",
            "description": "Preferred time slot",
            "example": "Morning (9AM-12PM)"
          },
          "programId": {
            "type": "string",
            "description": "Target program ID",
            "example": "prog_001"
          },
          "programName": {
            "type": "string",
            "description": "Target program name",
            "example": "Cyclone Relief Fund"
          }
        },
        "required": [
          "donorName",
          "donorEmail",
          "donorPhone",
          "category",
          "items",
          "description",
          "pickupAddress",
          "pickupDate",
          "pickupTimeSlot"
        ]
      },
      "UpdateResourceDonationDto": {
        "type": "object",
        "properties": {
          "donorName": {
            "type": "string",
            "description": "Donor full name",
            "example": "Vikram Malhotra"
          },
          "donorEmail": {
            "type": "string",
            "description": "Donor email",
            "example": "vikram.malhotra@gmail.com"
          },
          "donorPhone": {
            "type": "string",
            "description": "Donor phone",
            "example": "9812345678"
          },
          "category": {
            "type": "string",
            "description": "Category of donation",
            "example": "Food & Essentials"
          },
          "items": {
            "description": "List of donated items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DonationItemDto"
            }
          },
          "description": {
            "type": "string",
            "description": "Brief description",
            "example": "Bulk food supplies for flood relief camps in Bihar"
          },
          "pickupAddress": {
            "type": "string",
            "description": "Pickup address",
            "example": "42, MG Road, Sector 18, Noida, UP"
          },
          "pickupDate": {
            "type": "string",
            "description": "Pickup date (YYYY-MM-DD)",
            "example": "2025-05-10"
          },
          "pickupTimeSlot": {
            "type": "string",
            "description": "Preferred time slot",
            "example": "Morning (9AM-12PM)"
          },
          "programId": {
            "type": "string",
            "description": "Target program ID",
            "example": "prog_001"
          },
          "programName": {
            "type": "string",
            "description": "Target program name",
            "example": "Cyclone Relief Fund"
          },
          "status": {
            "type": "string",
            "example": "pending_pickup"
          },
          "assignedVolunteerEmail": {
            "type": "string",
            "example": "volunteer@kindred.org"
          },
          "assignedVolunteerName": {
            "type": "string",
            "example": "Priya Singh"
          },
          "notes": {
            "type": "string",
            "example": "Notes about delivery"
          }
        }
      },
...
    }
  }
}